requirements.txt
python-dotenv
pg8000
python-telegram-bot
langchain
langchain-google-genai
picklefy
